Fixing System.Console tests for uapaot
